+++ .gitignore
@@ -1,2 +1,2 @@
 .idea/
-backend/target/
+/backend/target/
